Move taintObject method to its own module to avoid circular references in taint-tracking-impl
