+++ packages/dd-trace/src/appsec/iast/taint-tracking/operations.js
@@ -2,11 +2,11 @@
 
 const TaintedUtils = require('@datadog/native-iast-taint-tracking')
 const { IAST_TRANSACTION_ID } = require('../iast-context')
-const iastLog = require('../iast-log')
 const iastTelemetry = require('../telemetry')
 const { REQUEST_TAINTED } = require('../telemetry/iast-metric')
 const { isInfoAllowed } = require('../telemetry/verbosity')
 const { getTaintTrackingImpl, getTaintTrackingNoop } = require('./taint-tracking-impl')
+const { taintObject } = require('./operations-taint-object')
 
 function createTransaction (id, iastContext) {
   if (id && iastContext) {
@@ -55,52 +55,8 @@ function newTaintedObject (iastContext, obj, name, type) {
   return result
 }
 
-function taintObject (iastContext, object, type, keyTainting, keyType) {
-  let result = object
-  const transactionId = iastContext?.[IAST_TRANSACTION_ID]
-  if (transactionId) {
-    const queue = [{ parent: null, property: null, value: object }]
-    const visited = new WeakSet()
-
-    while (queue.length > 0) {
-      const { parent, property, value, key } = queue.pop()
-      if (value === null) {
-        continue
-      }
-
-      try {
-        if (typeof value === 'string') {
-          const tainted = TaintedUtils.newTaintedString(transactionId, value, property, type)
-          if (!parent) {
-            result = tainted
-          } else if (keyTainting && key) {
-            const taintedProperty = TaintedUtils.newTaintedString(transactionId, key, property, keyType)
-            parent[taintedProperty] = tainted
-          } else {
-            parent[key] = tainted
-          }
-        } else if (typeof value === 'object' && !visited.has(value)) {
-          visited.add(value)
-
-          for (const key of Object.keys(value)) {
-            queue.push({ parent: value, property: property ? `${property}.${key}` : key, value: value[key], key })
-          }
-
-          if (parent && keyTainting && key) {
-            const taintedProperty = TaintedUtils.newTaintedString(transactionId, key, property, keyType)
-            parent[taintedProperty] = value
-          }
-        }
-      } catch (e) {
-        iastLog.error(`Error visiting property : ${property}`).errorAndPublish(e)
-      }
-    }
-  }
-  return result
-}
-
 function isTainted (iastContext, string) {
-  let result = false
+  let result
   const transactionId = iastContext?.[IAST_TRANSACTION_ID]
   if (transactionId) {
     result = TaintedUtils.isTainted(transactionId, string)
@@ -111,7 +67,7 @@ function isTainted (iastContext, string) {
 }
 
 function getRanges (iastContext, string) {
-  let result = []
+  let result
   const transactionId = iastContext?.[IAST_TRANSACTION_ID]
   if (transactionId) {
     result = TaintedUtils.getRanges(transactionId, string)
@@ -135,7 +91,7 @@ function enableTaintOperations (telemetryVerbosity) {
     onRemoveTransaction = onRemoveTransactionInformationTelemetry
   }
 
-  global._ddiast = getTaintTrackingImpl(telemetryVerbosity, taintObject, false)
+  global._ddiast = getTaintTrackingImpl(telemetryVerbosity, false)
 }
 
 function disableTaintOperations () {
